+++ src/margo.c
@@ -28,7 +28,7 @@
 struct mplex_key
 {
     hg_id_t id;
-    uint32_t mplex_id;
+    uint8_t mplex_id;
 };
 
 struct mplex_element
@@ -473,7 +473,7 @@ hg_id_t margo_register_name(margo_instance_id mid, const char *func_name,
 
 hg_id_t margo_register_name_mplex(margo_instance_id mid, const char *func_name,
     hg_proc_cb_t in_proc_cb, hg_proc_cb_t out_proc_cb, hg_rpc_cb_t rpc_cb,
-    uint32_t mplex_id, ABT_pool pool)
+    uint8_t mplex_id, ABT_pool pool)
 {
     struct mplex_key key;
     struct mplex_element *element;
@@ -513,7 +513,7 @@ hg_return_t margo_registered_name(margo_instance_id mid, const char *func_name,
 }
 
 hg_return_t margo_registered_name_mplex(margo_instance_id mid, const char *func_name,
-    uint32_t mplex_id, hg_id_t *id, hg_bool_t *flag)
+    uint8_t mplex_id, hg_id_t *id, hg_bool_t *flag)
 {
     hg_bool_t b;
     hg_return_t ret = margo_registered_name(mid, func_name, id, &b);
@@ -1006,7 +1006,7 @@ margo_instance_id margo_hg_info_get_instance(const struct hg_info *info)
 	return data->mid;
 }
 
-int margo_lookup_mplex(margo_instance_id mid, hg_id_t id, uint32_t mplex_id, ABT_pool *pool)
+int margo_lookup_mplex(margo_instance_id mid, hg_id_t id, uint8_t mplex_id, ABT_pool *pool)
 {
     struct mplex_key key;
     struct mplex_element *element;
@@ -1036,7 +1036,7 @@ int margo_lookup_mplex(margo_instance_id mid, hg_id_t id, uint32_t mplex_id, ABT
     return(0);
 }
 
-int margo_register_data_mplex(margo_instance_id mid, hg_id_t id, uint32_t mplex_id, void* data, void (*free_callback)(void *))
+int margo_register_data_mplex(margo_instance_id mid, hg_id_t id, uint8_t mplex_id, void* data, void (*free_callback)(void *))
 {
     struct mplex_key key;
     struct mplex_element *element;
@@ -1060,7 +1060,7 @@ int margo_register_data_mplex(margo_instance_id mid, hg_id_t id, uint32_t mplex_
     return(0);
 }
 
-void* margo_registered_data_mplex(margo_instance_id mid, hg_id_t id, uint32_t mplex_id)
+void* margo_registered_data_mplex(margo_instance_id mid, hg_id_t id, uint8_t mplex_id)
 {
     struct mplex_key key;
     struct mplex_element *element;
